% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoubleObs_package.r
\name{DoubleObs}
\alias{DoubleObs}
\title{Introduction}
\description{
DoubleObs provides two primary capabilities: 1) analyze hybrid double observer survey data in which some
of the animal groups are marked with a radio/satellite transmitter so it is known whether they were seen or not,
and 2) to simulate data from a hybrid double-observer survey and fit models to explore aspects of this type of analysis.
We describe these two capabilities below.
}
\details{
Analysis: Model Fitting/Abundance Estimation
Data
  There are a few requirements for the dataframe used in the analysis. Each record in the dataframe is for a single
  group of animals. The following 3 fields are required with specific names for each record:
   1) ch - must be a character string of length 3 with values of 0 or 1 in each position. The first character is 1 for groups that
   are marked (eg. an individual in group has a transmitter) and 0 otherwise.  The second character is a 1 if the front observer(s)
   detected the group and 0 otherwise.  The third character is the same for the rear observer(s).
   2) count - the count of animals in the group
   2) type - a factor variable with values "marked" and "unmarked".
  Any number of additional variables can be defined for use in the model fitting (eg lngs = log(count), vegetation type/cover etc).
Model Fitting
  Models are fitted with MARK using the RMark interface with the Huggins model primarily. As such, various steps need to be taken to use RMark and an
  understanding of RMark is helpful but not entirely necessary because the DoubleObs package contains functions: prep_data, prep_ddl and
  fit_models as wrappers for the RMark code. See the help files ?prep_data, ?prep_ddl and ?fit_models for more details. The prep_data
  and prep_ddl run the functions process.data and make.design.data in RMark and add some specific fields that are useful in fitting models to
  hybrid double observer survey data. Alternatively you can use the RMark functions separately and you can replace fit_models with calls to mark.wrapper and
  mark to fit models directly. The example ?hybrid_analysis shows different ways of fitting models.
Abundance estimation
  Two functions Nhat_group_marklist and Nhat_marklist in this package will construct abundance of groups and animals, respectively,
  for a set of models (marklist) and will compute the model averaged estimates as well.
While this package is mostly intended for double observer hybrid models, we have provided examples for a double observer survey without
marked groups ?double_observer_analysis and a sightability analysis in which a sample of marked animals is used to create the detection
model and those predictions are used for the survey data ?sightability_analysis.

Simulation:
The function simhet provides a simulation capability to generate hybrid double observer survey data with residual
heterogeneity. The function senerates simulation replicates of  double observer survey data with a known (marked) and unmarked portion of the population with
or without group sizes. Fits a sequence of models, and reports average abundance, average std error, confidence interval
coverage for each model and model average estimate.
}
\author{
Jeff Laake
}
