% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoubleObs_package.r
\docType{data}
\name{hybrid_analysis}
\alias{hybrid_analysis}
\alias{example_data}
\title{Example Double Observer Hybrid Data Analysis}
\format{
A data frame with 129 observations of 7 variables:
\describe{\item{ch}{a character vector containing the three character encounter history; first occasion indicates whether the group was marked or not}
\item{count}{number of animals in the group}
\item{het}{variable used to model heterogeneity (re-capture probability.}
\item{lngs}{ natural log of count}
\item{veg}{random vegetation variable}
\item{type}{"unmarked" or "marked"}
}
}
\description{
An example of double observer data with 25 marked (radio) groups and 150 other groups (no radio)
in the population. Data were generated with the following code using simhet function.
}
\details{
simres=simhet(1,150,25,beta=c(-.5,-.5,-1,2,.5),data=data.frame(veg=runif(175)),
             gen_formula="~-1+loc+veg+cov+lngs",use_gs=TRUE,gs_param=c(1,10),seed=9234397)
example_data=simres[[1]][[1]]$results[[1]]$data$data


The groups sizes were randomly generated from a gamma distribution with scale=10 and shape=1 (gs_param).
The true simulated population size is 175 groups and 1732 animals. A uniform random veg variable, and
unmodelled heterogeneity variable cov (uniform -1,1) and natural log of group size are used to
generate p for each location(loc).
}
\examples{
data(example_data)
# example using fit_models
fit_models(data=example_data,formula=c("veg+lngs","lngs","veg",""))
#example using RMark directly
dp=prep_data(example_data,model="Huggins")
# variables het, rear and loc are created in prep_ddl which call make.design.data in RMark
# and then creates other variables used to fit the models
ddl=prep_ddl(dp)
fit_do=function()
{
   # tag loss approach
   p.1=list(formula=~veg+lngs+loc+c:het:rear,share=TRUE)
   p.2=list(formula=~lngs+loc+c:het:rear,share=TRUE)
   p.3=list(formula=~veg+loc+c:het:rear,share=TRUE)
   p.4=list(formula=~loc+c:het:rear,share=TRUE)
   # hybrid approach
   p.5=list(formula=~veg+lngs+loc+type,share=TRUE)
   p.6=list(formula=~lngs+loc+type,share=TRUE)
   p.7=list(formula=~veg+loc+type,share=TRUE)
   p.8=list(formula=~loc+type,share=TRUE)
   cml=RMark::create.model.list("Huggins")
   results=RMark::mark.wrapper(cml,data=dp,ddl=ddl)
   return(results)
}
results=fit_do()
results
# Group abundance
Nhat_group_marklist(results,example_data)
# Group abundance estimate from MARK for best model
colSums(results[[1]]$results$derived[[1]])
# Group abundance not using 100 set (marked groups missed)
# This can be useful if not all marked missed groups can be found after survey
Nhat_group_marklist(results,example_data,marked_known=FALSE)
# MARK doesn't consider group size so their is no equivalent for total abundance
Nhat_marklist(results,example_data)
# Animal abundance not using 100 set (marked groups missed)
# This can be useful if not all marked missed groups can be found after survey
Nhat_marklist(results,example_data,marked_known=FALSE)
}
\keyword{datasets}
