% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nhat.r
\name{Nhat_grp}
\alias{Nhat_grp}
\alias{Nhat}
\alias{Nhat0}
\alias{Nhat_grp0}
\title{Abundance estimation for Huggins or HugFullHet models}
\usage{
Nhat_grp(mod,indices,data,known)

Nhat(mod,indices,data,known)

Nhat0(mod,indices,data,known)

Nhat_grp0(mod,indices,data,known)
}
\arguments{
\item{mod}{fitted Huggins or HugFullHet model for double observer data}

\item{indices}{model.index values of real parameters for p and pi(for Nhat0 and Nhat_grp0) from design data}

\item{data}{dataframe containing data used to fit model}

\item{known}{known marked animals (e.g. radio collar); either number of marked groups or sum of animals in marked groups}
}
\value{
list containing: 1) abundance estimate (either groups or animals) , 2) std error, 3) lower 95 percent confidence limit,
4) upper 95 percent confidence limit.
}
\description{
Computes estimate of group (Nhat_group) or animal (Nhat) abundance for Huggins model or
for HugFullHet with 2 mixtures in which pi is the proportion of animals with p=0.
}
\author{
Jeff Laake
}
