% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nhat_marklist.r
\name{Nhat_marklist}
\alias{Nhat_marklist}
\alias{Nhat_group_marklist}
\alias{mod_avgN}
\title{Abundance estimation for Huggins or HugFullHet models from marklist of fitted models}
\usage{
Nhat_group_marklist(marklist,data,marked_known=TRUE)
Nhat_marklist(marklist,data,marked_known=TRUE)
mod_avgN(marklist,Nmatrix,Ngrp=TRUE,data)
}
\arguments{
\item{marklist}{list of fitted mark models}

\item{data}{dataframe containing data used to fit model or subset of that data}

\item{marked_known}{if FALSE includes marked groups into estimation,  otherwise adds number in marked groups to abundance}

\item{Ngrp}{if TRUE, model average is for group abundance}

\item{Nmatrix}{matrix of estimates in order of model.table (return value from Nhat_group_marklist and Mhat_marklist)}
}
\value{
matrix of group abundance estimates (Nhat_group_marklist) or animal abundance (Nhat_marklist) with std errors, and confidence intervals for each model and model averaged values
}
\description{
Uses (Nhat_group) and/or animal (Nhat) abundance to estimate abundance for a list
of fitted mark models. Works with results of fit_models
}
\author{
Jeff Laake
}
